+++ joylive-plugin/joylive-registry/joylive-registry-springcloud2.1/src/main/resources/META-INF/services/com.jd.live.agent.core.plugin.definition.PluginDefinition
@@ -1,4 +1,4 @@
 com.jd.live.agent.plugin.registry.springcloud.v2_1.definition.RegistryDefinition
 com.jd.live.agent.plugin.registry.springcloud.v2_1.definition.FeignClientFactoryBeanDefinition
-com.jd.live.agent.plugin.registry.springcloud.v2_1.definition.DiscoveryClientDefinition
+com.jd.live.agent.plugin.registry.springcloud.v2_1.definition.DiscoveryClientInstanceSupplierDefinition
 com.jd.live.agent.plugin.registry.springcloud.v2_1.definition.SimpleDiscoveryClientDefinition
