+++ joylive-plugin/joylive-registry/joylive-registry-springcloud2.2/src/main/java/com/jd/live/agent/plugin/registry/springcloud/v2_2/interceptor/DiscoveryInstanceSupplierConstructorInterceptor.java
@@ -25,13 +25,13 @@ import org.springframework.cloud.loadbalancer.core.ServiceInstanceListSupplier;
 /**
  * DiscoveryClientConstructorInterceptor
  */
-public class DiscoveryClientConstructorInterceptor extends InterceptorAdaptor {
+public class DiscoveryInstanceSupplierConstructorInterceptor extends InterceptorAdaptor {
 
-    private static final Logger logger = LoggerFactory.getLogger(DiscoveryClientConstructorInterceptor.class);
+    private static final Logger logger = LoggerFactory.getLogger(DiscoveryInstanceSupplierConstructorInterceptor.class);
 
     private final Registry registry;
 
-    public DiscoveryClientConstructorInterceptor(Registry registry) {
+    public DiscoveryInstanceSupplierConstructorInterceptor(Registry registry) {
         this.registry = registry;
     }
